gateway/requirements.txt
fastapi==0.89.1
uvicorn==0.20.0
requests==2.32.3
PyJWT==2.10.1
cryptography==44.0.2
